Converted main program to use classes
Game elements are randomly positioned

+++ using_classes_and_objects/GameElement.pde
@@ -2,15 +2,15 @@ class GameElement {
   
   // properties (variables) accessible only inside an instance of this class go here
   
-  int x;      // horizontal position of game element
-  int y;      // vertical position of game element
+  float x;      // horizontal position of game element
+  float y;      // vertical position of game element
   int size;   // width and height of game element
   int shade;  // grayscale value / shade value of grey for element
   int points; // points earned for clicking this element
   
   // constructor (this ones runs when an instance of this object is created, like "setup"
   //              in main program)
-  GameElement(int _x, int _y, int _size, int _shade, int _points) {
+  GameElement(float _x, float _y, int _size, int _shade, int _points) {
     
     // Take all the values passed in the parameters, and assign to the
     // properties (variables) that belong to this instance of the class
